Reorganize sources: ipc -> sock
